requirements.txt
Django==1.9.5
django-countries==3.4.1
django-geoposition==0.2.3
django-localflavor==1.3
django-nested-admin==3.0.8
django-nested-inline==0.3.6
Pillow==3.3.1
psycopg2==2.6.2
python-monkey-business==1.0.0
six==1.10.0
